Move caching into own module
