+++ lib/rp.js
@@ -8,11 +8,8 @@ var Bluebird = require('bluebird'),
     isString = require('lodash/isString'),
     isUndefined = require('lodash/isUndefined'),
     keys = require('lodash/keys'),
-    errors = require('./errors.js');
-
-// Used to handle HTTP 1.1 Cache
-var moment = require('moment'),
-    Cache = require('memory-cache');
+    errors = require('./errors.js'),
+    httpCache = require('./httpcache');
 
 
 // Load Request freshly - so that users can require an unaltered request instance!
@@ -43,15 +40,6 @@ var defaultTransformations = {
     }
 };
 
-var _computeAgeOfEntry = function (entry, now) {
-
-  var created_at = moment(entry.created_at);
-  var age = moment.duration(now.diff(entry.created_at)).asSeconds();
-
-  // normalize the age in seconds
-  age = -Math.floor(-age);
-  entry.content.headers.age = age;
-};
 
 function RP$callback(err, response, body) {
 
@@ -71,15 +59,16 @@ function RP$callback(err, response, body) {
 
     // 304 response means we can reuse the cache entry we already had
     if (response && response.statusCode === 304 && self._rp_options.cache) {
-        var now = moment();
-
-        var cacheEntry = Cache.get('GET:' + response.request.uri.href);
+        
+        var now = httpCache.now(),
+            cacheEntry = httpCache.get(response.request.uri.href);
+        
         if (!cacheEntry) {
             throw new Error(response.request.uri.href + ' should have exit in the cache but it is not there');
         }
 
         cacheEntry.created_at = now;
-        _computeAgeOfEntry(cacheEntry, now);
+        httpCache.computeAgeOfEntry(cacheEntry, now);
         response = cacheEntry.content;
     }
 
@@ -106,8 +95,6 @@ function RP$callback(err, response, body) {
         }
 
     } else {
-
-
 
         if (isFunction(self._rp_options.transform)) {
 
@@ -136,103 +123,6 @@ function RP$callback(err, response, body) {
 }
 
 
-/**
- * HTTP 1.1 Cache handling
- */
-// Cache.debug(true);
-
-var LAST_MODIFIED_FORMAT = 'ddd, DD MMM YYYY HH:mm:ss z';
-var isResolveWithFullResponseSet = false;
-
-
-var _useInMemoryCache = function (requestOpts) {
-    return !isUndefined(requestOpts.cache) && requestOpts.cache === true && (isUndefined(requestOpts.method) || requestOpts.method.toUpperCase() === 'GET');
-};
-
-
-
-var _httpCacheHandler = function (response) {
-
-    // Restore option to its original state
-    this._rp_options.resolveWithFullResponse = isResolveWithFullResponseSet;
-
-    var useCache = true;
-    var headers = response.headers;
-    var cacheKey = 'GET:' + response.request.href;
-
-    // Don't cache if Cache-Control doesn't exist
-    if (isUndefined(headers['cache-control'])) {
-        useCache = false;
-    }
-
-    // Don't cache on Cache-Control.no-cache
-    if (!isUndefined(headers['cache-control']) && /no-cache/.test(headers['cache-control'])) {
-        useCache = false;
-    }
-
-    // Don't cache if Cache-Control.max-age doesn't exist
-    if (!isUndefined(headers['cache-control']) && !/max-age=(\d+)/.test(headers['cache-control'])) {
-        useCache = false;
-    }
-
-    // Don't cache if Cache-Control.max-age is 0
-    if (!isUndefined(headers['cache-control']) && /max-age=0/.test(headers['cache-control'])) {
-        useCache = false;
-    }
-
-    if (useCache) {
-
-        var age = !isUndefined(headers['age']) ? headers['age'] : 0,
-            maxAge = parseInt(headers['cache-control'].match(/max-age=(\d+)/)[1]),
-            lastModified;
-
-        // Get Last-Modified
-        if (isUndefined(headers['last-modified'])) {
-            lastModified = headers['last-modified'];
-        }
-
-        if (maxAge > 0) {
-            var now = moment();
-            var createdAt = moment(now).subtract(age, 's');  // we reduce created_at from age to have a
-                                                             // consistent age with the backend
-            var value = {
-                max_age: maxAge,
-                content: response,
-                created_at: createdAt,
-                last_modified: lastModified || createdAt
-            };
-
-            // Add/Update entry age
-            _computeAgeOfEntry(value, now);
-
-            // This insert will also refresh `created_at` in case of 304 responses
-            Cache.put(cacheKey, value);
-        }
-    }
-
-    // Add age if undefined
-    response.headers.age = headers['age'] || 0;
-
-    // @TODO This return is being ignored
-    return isResolveWithFullResponseSet ? response : response.body;
-};
-
-
-var _isCacheStale = function (cachedResponse) {
-
-    if (cachedResponse) {
-        var now = moment();
-        _computeAgeOfEntry(cachedResponse, now);
-
-        // Check if cache entry is a stale
-        if (cachedResponse.content.headers.age > cachedResponse.max_age) {
-            return cachedResponse.last_modified.format(LAST_MODIFIED_FORMAT);
-        }
-    }
-
-    return false;
-};
-
 //
 // Init
 //
@@ -245,29 +135,28 @@ request.Request.prototype.init = function RP$initInterceptor(options) {
     // Init may be called again - currently in case of redirects
     if (isPlainObject(options) && isUndefined(self._callback) && isUndefined(self._rp_promise)) {
 
-        var cachedResponse = null;
-
-        if (_useInMemoryCache(options)) {
-            cachedResponse = Cache.get('GET:' + options.uri);
+        if (httpCache.useCache(options)) {
+            var cachedResponse = httpCache.get(options.uri);
 
             // Store current status of `resolveWithFullResponse`
-            isResolveWithFullResponseSet = options.resolveWithFullResponse || false;
+            httpCache.isResolveWithFullResponseSet = options.resolveWithFullResponse || false;
 
             // Force resolveWithFullResponse to have access to the response header
             options.resolveWithFullResponse = true;
 
             if (cachedResponse) {
-                var lastModified = _isCacheStale(cachedResponse);
+                var lastModified = httpCache.isStale(cachedResponse);
 
                 // Return cached resource if not stale
                 if (!lastModified) {
-                    _computeAgeOfEntry(cachedResponse, moment());
+                    httpCache.computeAgeOfEntry(cachedResponse);
 
                     self._rp_promise = new Bluebird(function (resolve) {
-                      resolve(isResolveWithFullResponseSet ? cachedResponse.content : cachedResponse.content.body);
+                      resolve(httpCache.isResolveWithFullResponseSet ? cachedResponse.content : cachedResponse.content.body);
                     });
 
                     return;
+
                 } else {
                     // Add pre-flight header
                     options.headers = options.headers || {};
@@ -275,7 +164,8 @@ request.Request.prototype.init = function RP$initInterceptor(options) {
 
                     var etag = cachedResponse.content && cachedResponse.content.headers &&
                         cachedResponse.content.headers['etag'];
-                    if ( etag ) {
+                    
+                    if (etag) {
                         options.headers['If-None-Match'] = etag;
                     }
                 }
@@ -288,8 +178,8 @@ request.Request.prototype.init = function RP$initInterceptor(options) {
                 self._rp_reject = reject;
             });
 
-            if (_useInMemoryCache(options)) {
-                self._rp_promise.then(_httpCacheHandler.bind(self));
+            if (httpCache.useCache(options)) {
+                self._rp_promise.then(httpCache.responseHandler.bind(self));
             }
         }
